+++ Adjust/ADJActivityState.h
@@ -16,7 +16,7 @@
 @property (nonatomic, assign) BOOL askingAttribution;
 @property (nonatomic, assign) BOOL isThirdPartySharingDisabled;
 
-@property (nonatomic, copy) NSString *uuid;
+@property (nonatomic, copy) NSString *dedupeToken;
 @property (nonatomic, copy) NSString *deviceToken;
 @property (nonatomic, assign) BOOL updatePackages;