+++ Adjust/ADJLinkResolution.h
@@ -1,0 +1,17 @@
+//
+//  ADJLinkResolution.h
+//  Adjust
+//
+//  Created by Pedro S. on 26.04.21.
+//  Copyright © 2021 adjust GmbH. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface ADJLinkResolution : NSObject
+
++ (void)resolveLinkWithUrl:(nonnull NSURL *)url
+     resolveUrlSuffixArray:(nullable NSArray<NSString *> *)resolveUrlSuffixArray
+                  callback:(nonnull void (^)(NSURL *_Nullable resolvedLink))callback;
+
+@end