+++ examples/AdjustExample-FbPixel/AdjustExample-FbPixel/WKWebViewController.m
@@ -27,18 +27,23 @@
 }
 
 - (void)loadWKWebView {
-    WKWebView *webView = [[NSClassFromString(@"WKWebView") alloc] initWithFrame:self.view.bounds];
-    webView.navigationDelegate = self;
-    [self.view addSubview:webView];
+    WKWebView *wkWebView = [[NSClassFromString(@"WKWebView") alloc] initWithFrame:self.view.bounds];
+    wkWebView.navigationDelegate = self;
+    [self.view addSubview:wkWebView];
 
     _adjustBridge = [[AdjustBridge alloc] init];
-    [_adjustBridge loadWKWebViewBridge:webView wkWebViewDelegate:self];
+    [_adjustBridge loadWKWebViewBridge:wkWebView wkWebViewDelegate:self];
     [_adjustBridge augmentHybridWebView];
+    
+    _jsContext = [wkWebView valueForKeyPath:@"documentView.webView.mainFrame.javaScriptContext"];
+    _jsContext[@"console"][@"log"] = ^(JSValue * msg) {
+        NSLog(@"JavaScript %@ log message: %@", [JSContext currentContext], msg);
+    };
 
     NSString *htmlPath = [[NSBundle mainBundle] pathForResource:@"AdjustExample-FbPixel" ofType:@"html"];
     NSString *appHtml = [NSString stringWithContentsOfFile:htmlPath encoding:NSUTF8StringEncoding error:nil];
     NSURL *baseURL = [NSURL fileURLWithPath:htmlPath];
-    [webView loadHTMLString:appHtml baseURL:baseURL];
+    [wkWebView loadHTMLString:appHtml baseURL:baseURL];
 }
 
 - (void)callWkHandler:(id)sender {